Replace bare flag with bare file
