+++ bare.ts
@@ -1,0 +1,24 @@
+import { Environment } from "./src/environment.ts";
+import { FileLoader, Loader } from "./src/loader.ts";
+
+export interface Options {
+  includes?: string | Loader;
+  autoDataVarname?: boolean;
+  dataVarname?: string;
+  autoescape?: boolean;
+}
+
+export default function (options: Options = {}): Environment {
+  const loader = typeof options.includes === "object"
+    ? options.includes
+    : new FileLoader(options.includes || Deno.cwd());
+
+  const env = new Environment({
+    loader,
+    dataVarname: options.dataVarname || "it",
+    autoescape: options.autoescape ?? false,
+    autoDataVarname: options.autoDataVarname ?? true,
+  });
+
+  return env;
+}
